services/memory-gateway/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
requests
google-cloud-run
pydantic
